+++ plugin/src/main/java/com/jacobibanez/plugin/android/godotplaygameservices/leaderboards/LeaderboardMapper.kt
@@ -4,47 +4,66 @@ import com.google.android.gms.games.leaderboard.Leaderboard
 import com.google.android.gms.games.leaderboard.LeaderboardScore
 import com.google.android.gms.games.leaderboard.LeaderboardVariant
 import com.jacobibanez.plugin.android.godotplaygameservices.players.fromPlayer
+import com.jacobibanez.plugin.android.godotplaygameservices.utils.toStringAndSave
 import org.godotengine.godot.Dictionary
+import org.godotengine.godot.Godot
 
 /** @suppress */
-fun fromLeaderboardScore(score: LeaderboardScore): Dictionary = Dictionary().apply {
-    put("displayRank", score.displayRank)
-    put("displayScore", score.displayScore)
-    put("rank", score.rank)
-    put("rawScore", score.rawScore)
+fun fromLeaderboardScore(godot: Godot, score: LeaderboardScore): Dictionary =
+    Dictionary().apply {
+        put("displayRank", score.displayRank)
+        put("displayScore", score.displayScore)
+        put("rank", score.rank)
+        put("rawScore", score.rawScore)
+        put("scoreHolderDisplayName", score.scoreHolderDisplayName)
+        put("scoreTag", score.scoreTag)
+        put("timestampMillis", score.timestampMillis)
+        score.scoreHolder?.let { player ->
+            put("scoreHolder", fromPlayer(godot, player))
 
-    score.scoreHolder?.let {
-        put("scoreHolder", fromPlayer(it))
+            put(
+                "scoreHolderHiResImageUri",
+                score.scoreHolderHiResImageUri.toStringAndSave(
+                    godot,
+                    "scoreHolderHiResImageUri",
+                    player.playerId
+                )
+            )
+            put(
+                "scoreHolderIconImageUri",
+                score.scoreHolderIconImageUri.toStringAndSave(
+                    godot,
+                    "scoreHolderIconImageUri",
+                    player.playerId
+                )
+            )
+        }
     }
 
-    put("scoreHolderDisplayName", score.scoreHolderDisplayName)
-    put("scoreHolderHiResImageUri", score.scoreHolderHiResImageUri?.toString())
-    put("scoreHolderIconImageUri", score.scoreHolderIconImageUri?.toString())
-    put("scoreTag", score.scoreTag)
-    put("timestampMillis", score.timestampMillis)
-}
-
 /** @suppress */
-fun fromLeaderboard(leaderboard: Leaderboard): Dictionary = Dictionary().apply {
+fun fromLeaderboard(godot: Godot, leaderboard: Leaderboard): Dictionary = Dictionary().apply {
     put("leaderboardId", leaderboard.leaderboardId)
     put("displayName", leaderboard.displayName)
-    put("iconImageUri", leaderboard.iconImageUri?.toString())
-    ScoreOrder.fromOrder(leaderboard.scoreOrder)?.let {
-        put("scoreOrder", it.name)
-    }
     put("variants", fromLeaderboardVariant(leaderboard.variants))
+    ScoreOrder.fromOrder(leaderboard.scoreOrder)?.let { put("scoreOrder", it.name) }
+    leaderboard.iconImageUri?.let {
+        put(
+            "iconImageUri",
+            it.toStringAndSave(
+                godot,
+                "iconImageUri",
+                leaderboard.leaderboardId
+            )
+        )
+    }
 }
 
 /** @suppress */
 fun fromLeaderboardVariant(variants: ArrayList<LeaderboardVariant>): List<Dictionary> = variants
     .map { variant ->
         Dictionary().apply {
-            Collection.fromType(variant.collection)?.let {
-                put("collection", it.name)
-            }
-            TimeSpan.fromSpan(variant.timeSpan)?.let {
-                put("timeSpan", it.name)
-            }
+            Collection.fromType(variant.collection)?.let { put("collection", it.name) }
+            TimeSpan.fromSpan(variant.timeSpan)?.let { put("timeSpan", it.name) }
             put("displayPlayerRank", variant.displayPlayerRank)
             put("displayPlayerScore", variant.displayPlayerScore)
             put("numScores", variant.numScores)