+++ plugin/src/main/java/com/jacobibanez/plugin/android/godotplaygameservices/players/PlayerMapper.kt
@@ -3,23 +3,19 @@ package com.jacobibanez.plugin.android.godotplaygameservices.players
 import com.google.android.gms.games.Player
 import com.google.android.gms.games.PlayerLevel
 import com.google.android.gms.games.PlayerLevelInfo
+import com.jacobibanez.plugin.android.godotplaygameservices.utils.toStringAndSave
 import org.godotengine.godot.Dictionary
+import org.godotengine.godot.Godot
 
 /** @suppress */
-fun fromPlayer(player: Player): Dictionary = Dictionary().apply {
-    put("bannerImageLandscapeUri", player.bannerImageLandscapeUri?.toString())
-    put("bannerImagePortraitUri", player.bannerImagePortraitUri?.toString())
+fun fromPlayer(godot: Godot, player: Player): Dictionary = Dictionary().apply {
     put("displayName", player.displayName)
-    put("hiResImageUri", player.hiResImageUri?.toString())
-    put("iconImageUri", player.iconImageUri?.toString())
     put("playerId", player.playerId)
     put("retrievedTimestamp", player.retrievedTimestamp)
     put("title", player.title)
     put("hasHiResImage", player.hasHiResImage())
     put("hasIconImage", player.hasIconImage())
-    player.levelInfo?.let {
-        put("levelInfo", fromPlayerLevelInfo(it))
-    }
+    player.levelInfo?.let { put("levelInfo", fromPlayerLevelInfo(it)) }
     player.currentPlayerInfo?.let { currentPlayerInfo ->
         FriendsListVisibilityStatus.fromStatus(currentPlayerInfo.friendsListVisibilityStatus)
             ?.let {
@@ -30,6 +26,46 @@ fun fromPlayer(player: Player): Dictionary = Dictionary().apply {
         PlayerFriendStatus.fromStatus(relationshipInfo.friendStatus)?.let {
             put("friendStatus", it.name)
         }
+    }
+    player.bannerImageLandscapeUri?.let {
+        put(
+            "bannerImageLandscapeUri",
+            it.toStringAndSave(
+                godot,
+                "bannerImageLandscapeUri",
+                player.playerId
+            )
+        )
+    }
+    player.bannerImagePortraitUri?.let {
+        put(
+            "bannerImagePortraitUri",
+            it.toStringAndSave(
+                godot,
+                "bannerImagePortraitUri",
+                player.playerId
+            )
+        )
+    }
+    player.hiResImageUri?.let {
+        put(
+            "hiResImageUri",
+            it.toStringAndSave(
+                godot,
+                "hiResImageUri",
+                player.playerId
+            )
+        )
+    }
+    player.iconImageUri?.let {
+        put(
+            "iconImageUri",
+            it.toStringAndSave(
+                godot,
+                "iconImageUri",
+                player.playerId
+            )
+        )
     }
 }